backend/requirements.txt
fastapi>=0.90.0,<0.111.0 # Use a specific range or just the library names
uvicorn[standard]>=0.18.0,<0.28.0
requests>=2.28.0,<3.0.0
PyYAML>=6.0,<7.0